requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
shap>=0.49.1
matplotlib>=3.8.0
pyyaml>=6.0.0
joblib>=1.3.0
pytest>=8.0.0
pre-commit>=3.6.0